refactor(application/web): web 패키지 내 spring, spark 패키지 생성
